Rename threading backend files. Delete useless files.
