+++ arbor/threadinfo.cpp
@@ -1,13 +1,4 @@
 #include <string>
 
-#include <arbor/threadinfo.hpp>
 
-#include "threading/threading.hpp"
-
-namespace arb {
-
-std::string thread_implementation() {
-    return threading::description();
-}
-
-} // namespace arb
+#include "threading/thread_info.hpp"
