--- arbor/threading/cthread.hpp
@@ -1,7 +1,0 @@
-#pragma once
-
-// task_group definition
-#include "cthread_impl.hpp"
-
-// and sorts use cthread_parallel_stable_sort
-#include "cthread_sort.hpp"
