+++ arbor/threading/threading.cpp
@@ -1,76 +1,143 @@
-#include <cstdlib>
+#include <atomic>
+#include <cassert>
+#include <cstring>
 #include <exception>
+#include <iostream>
 #include <regex>
-#include <string>
-
-#include <arbor/arbexcept.hpp>
-#include <arbor/util/optional.hpp>
-#include <hardware/node_info.hpp>
 
 #include "threading.hpp"
-#include "util/strprintf.hpp"
+#include "thread_info.hpp"
+#include "arbor/execution_context.hpp"
 
-namespace arb {
-namespace threading {
+using namespace arb::threading::impl;
+using namespace arb::threading;
+using namespace arb;
 
-// Test environment variables for user-specified count of threads.
-//
-// ARB_NUM_THREADS is used if set, otherwise OMP_NUM_THREADS is used.
-//
-// If neither variable is set, returns no value.
-//
-// Valid values for the environment variable are:
-//  0 : Arbor is responsible for picking the number of threads.
-//  >0: The number of threads to use.
-//
-// Throws std::runtime_error:
-//  ARB_NUM_THREADS or OMP_NUM_THREADS is set with invalid value.
-util::optional<size_t> get_env_num_threads() {
-    const char* str;
-
-    // select variable to use:
-    //   If ARB_NUM_THREADS_VAR is set, use $ARB_NUM_THREADS_VAR
-    //   else if ARB_NUM_THREAD set, use it
-    //   else if OMP_NUM_THREADS set, use it
-    if (auto nthreads_var_name = std::getenv("ARB_NUM_THREADS_VAR")) {
-        str = std::getenv(nthreads_var_name);
+task notification_queue::try_pop() {
+    task tsk;
+    lock q_lock{q_mutex_, std::try_to_lock};
+    if (q_lock && !q_tasks_.empty()) {
+        tsk = std::move(q_tasks_.front());
+        q_tasks_.pop_front();
     }
-    else if (! (str = std::getenv("ARB_NUM_THREADS"))) {
-        str = std::getenv("OMP_NUM_THREADS");
+    return tsk;
+}
+
+task notification_queue::pop() {
+    task tsk;
+    lock q_lock{q_mutex_};
+    while (q_tasks_.empty() && !quit_) {
+        q_tasks_available_.wait(q_lock);
     }
-
-    // If the selected var is unset set the number of threads to
-    // the hint given by the standard library
-    if (!str) {
-        return util::nullopt;
+    if (!q_tasks_.empty()) {
+        tsk = std::move(q_tasks_.front());
+        q_tasks_.pop_front();
     }
+    return tsk;
+}
 
-    auto nthreads = std::strtoul(str, nullptr, 10);
-
-    // check that the environment variable string describes a non-negative integer
-    if (errno==ERANGE ||
-        !std::regex_match(str, std::regex("\\s*\\d*[0-9]\\d*\\s*")))
+bool notification_queue::try_push(task& tsk) {
     {
-        throw arbor_exception(util::pprintf(
-            "requested number of threads \"{}\" is not a valid value", str));
+        lock q_lock{q_mutex_, std::try_to_lock};
+        if (!q_lock) return false;
+        q_tasks_.push_back(std::move(tsk));
+        tsk = 0;
     }
-
-    return nthreads;
+    q_tasks_available_.notify_all();
+    return true;
 }
 
-std::size_t num_threads_init() {
-    std::size_t n = 0;
-
-    if (auto env_threads = get_env_num_threads()) {
-        n = env_threads.value();
+void notification_queue::push(task&& tsk) {
+    {
+        lock q_lock{q_mutex_};
+        q_tasks_.push_back(std::move(tsk));
     }
-
-    if (!n) {
-        n = hw::node_processors();
-    }
-
-    return n? n: 1;
+    q_tasks_available_.notify_all();
 }
 
-} // namespace threading
-} // namespace arb
+void notification_queue::quit() {
+    {
+        lock q_lock{q_mutex_};
+        quit_ = true;
+    }
+    q_tasks_available_.notify_all();
+}
+
+void task_system::run_tasks_loop(int i){
+    while (true) {
+        task tsk;
+        for (unsigned n = 0; n != count_; n++) {
+            tsk = q_[(i + n) % count_].try_pop();
+            if (tsk) break;
+        }
+        if (!tsk) tsk = q_[i].pop();
+        if (!tsk) break;
+        tsk();
+    }
+}
+
+void task_system::try_run_task() {
+    auto nthreads = get_num_threads();
+    task tsk;
+    for (int n = 0; n != nthreads; n++) {
+        tsk = q_[n % nthreads].try_pop();
+        if (tsk) {
+            tsk();
+            break;
+        }
+    }
+}
+
+task_system::task_system() : count_(num_threads_init()), q_(num_threads_init()) {
+    assert( nthreads > 0);
+
+    // now for the main thread
+    auto tid = std::this_thread::get_id();
+    thread_ids_[tid] = 0;
+
+    for (unsigned i = 1; i < count_; i++) {
+        threads_.emplace_back([this, i]{run_tasks_loop(i);});
+        tid = threads_.back().get_id();
+        thread_ids_[tid] = i;
+    }
+}
+
+task_system::task_system(int nthreads) : count_(nthreads), q_(nthreads) {
+    assert( nthreads > 0);
+
+    // now for the main thread
+    auto tid = std::this_thread::get_id();
+    thread_ids_[tid] = 0;
+
+    for (unsigned i = 1; i < count_; i++) {
+        threads_.emplace_back([this, i]{run_tasks_loop(i);});
+        tid = threads_.back().get_id();
+        thread_ids_[tid] = i;
+    }
+}
+
+task_system::~task_system() {
+    for (auto& e: q_) e.quit();
+    for (auto& e: threads_) e.join();
+}
+
+void task_system::async(task tsk) {
+    auto i = index_++;
+
+    for (unsigned n = 0; n != count_; n++) {
+        if (q_[(i + n) % count_].try_push(tsk)) return;
+    }
+    q_[i % count_].push(std::move(tsk));
+}
+
+int task_system::get_num_threads() {
+    return threads_.size() + 1;
+}
+
+std::unordered_map<std::thread::id, std::size_t> task_system::get_thread_ids() {
+    return thread_ids_;
+};
+
+task_system_handle arb::make_thread_pool() {
+    return task_system_handle(new task_system(num_threads_init()));
+}
