+++ example/brunel/brunel_miniapp.cpp
@@ -15,7 +15,6 @@
 #include <arbor/profile/profiler.hpp>
 #include <arbor/recipe.hpp>
 #include <arbor/simulation.hpp>
-#include <arbor/threadinfo.hpp>
 #include <arbor/version.hpp>
 
 #include <aux/ioutil.hpp>
@@ -298,8 +297,7 @@ void banner(proc_allocation nd, const execution_context* ctx) {
     std::cout << "  Arbor miniapp\n";
     std::cout << "  - distributed : " << ctx->distributed->size()
               << " (" << ctx->distributed->name() << ")\n";
-    std::cout << "  - threads     : " << nd.num_threads
-              << " (" << arb::thread_implementation() << ")\n";
+    std::cout << "  - threads     : " << nd.num_threads << "\n";
     std::cout << "  - gpus        : " << nd.num_gpus << "\n";
     std::cout << "==========================================\n";
 }
